% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess-package.R
\docType{package}
\name{loess-package}
\alias{loess}
\alias{loess-package}
\title{This package is designed to teach Loess regression.}
\description{
This package is designed to help the teaching of Loess regression. The primary function, loess_vis, creates a ggplot2 image whereby the points are shaded based upon their weight when estimating y at a given x value. The accompanying Shiny application utilizes this function by allowing the user to modify the parameters of the loess_vis function.
}
\author{
\strong{Maintainer}: Jason Bryer \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{loess}
\keyword{regression}
\keyword{visualization}
