% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{loess_shiny}
\alias{loess_shiny}
\title{Run the Loess shiny app.}
\usage{
loess_shiny(span_range = c(0.05, 1), ...)
}
\arguments{
\item{span_range}{the range of values the user is allowed to set the span.}

\item{...}{not currently used.}
}
\description{
This will start the shiny app to demonstrate the loess_vis function.
}
