% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess_vis.R
\name{loess_vis}
\alias{loess_vis}
\title{Visualization designed to teach Loess regression.}
\usage{
loess_vis(
  formula,
  data,
  center = median(data[, all.vars(formula)[2], drop = TRUE]),
  degree = 2,
  span = 0.3,
  draw_loess = FALSE,
  show_loess = TRUE,
  xlab = all.vars(formula)[2],
  ylab = all.vars(formula)[1],
  ...
)
}
\arguments{
\item{formula}{formula used for the regression.}

\item{data}{the data.frame to run the regression.}

\item{center}{the point where the Loess regression is being estimated.}

\item{degree}{the degree of the polynomials to be used, normally 1 or 2.
See \code{\link[stats:loess]{stats::loess()}} for more information.}

\item{span}{the parameter α which controls the degree of smoothing.}

\item{draw_loess}{draw the Loess regression up to \code{x <= center}.}

\item{show_loess}{draw the full Loess regression line.}

\item{xlab}{label for the x-axis.}

\item{ylab}{lable for the y-axis.}

\item{...}{currently unused.}
}
\description{
This function is designed to display a scatterplot where the points are
shaded based upon their weight for estimating the Loess regression line
at a given x value.
}
\examples{
data("faithful")
loess::loess_vis(formula = eruptions ~ waiting, data = faithful)
}
